+++ CHANGELOG.md
@@ -3,3 +3,6 @@
         * Move extension code to src
         * Build **extensions/chrome** and **extensions/firefox**
             * Firefox only supports manifest v2, and Chrome supports manifest v3
+
+* 02/16/2024
+    * Fix inspecting async render and compute pipelines